+++ forgerock-authenticator/src/main/java/org/forgerock/android/auth/PushNotification.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020 - 2021 ForgeRock. All rights reserved.
+ * Copyright (c) 2020 - 2022 ForgeRock. All rights reserved.
  *
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -7,9 +7,17 @@
 
 package org.forgerock.android.auth;
 
-import androidx.annotation.NonNull;
+import android.os.Build;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.biometric.BiometricPrompt;
+
+import org.forgerock.android.auth.biometric.BiometricAuth;
+import org.forgerock.android.auth.biometric.BiometricAuthCompletionHandler;
 import org.forgerock.android.auth.exception.InvalidNotificationException;
+import org.forgerock.android.auth.exception.PushMechanismException;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -23,15 +31,15 @@ import java.util.TimeZone;
 public class PushNotification extends ModelObject<PushNotification> {
 
     /** Unique identifier for PushNotification object associated with the mechanism */
-    private String id;
+    private final String id;
     /** Unique identifier of the Mechanism associated with this PushNotification */
     private final String mechanismUID;
     /** Message Identifier that was received with this notification */
-    private String messageId;
+    private final String messageId;
     /** Base64 challenge that was sent with this notification */
-    private String challenge;
+    private final String challenge;
     /** The AM load balance cookie */
-    private String amlbCookie;
+    private final String amlbCookie;
     /** Date that the notification was received */
     private final Calendar timeAdded;
     /** Date that the notification has expired */
@@ -42,15 +50,26 @@ public class PushNotification extends ModelObject<PushNotification> {
     private boolean approved;
     /** Determines if the PushNotification has been interacted with the user. */
     private boolean pending;
+    /** The JSON String containing the custom attributes added to this notification */
+    private final String customPayload;
+    /** Message that was received with this notification */
+    private final String message;
     /** The mechanism associated with this notification **/
     private PushMechanism pushMechanism;
+    /** The type of push notification **/
+    private final PushType pushType;
+    /** The numbers used in the push challenge **/
+    private final String numbersChallenge;
+    /** The context information to this notification. */
+    private final String contextInfo;
 
     private static final String TAG = PushNotification.class.getSimpleName();
 
     /**
      * Creates the PushNotification object with given data
      * @param mechanismUID Mechanism UUID associated with the PushNotification
-     * @param messageId message identifier from the message payload
+     * @param messageId message identifier from the notification payload
+     * @param message message from the notification payload
      * @param challenge challenge from message payload
      * @param amlbCookie load balance cookie from OpenAM
      * @param timeAdded Date when the notification is delivered
@@ -58,13 +77,19 @@ public class PushNotification extends ModelObject<PushNotification> {
      * @param ttl time-to-live value from message payload
      * @param approved boolean indicator of whether notification is still pending or not
      * @param pending boolean indicator of whether notification is approved or not
+     * @param customPayload JSON String containing the custom attributes
+     * @param numbersChallenge numbers used in the push challenge
+     * @param contextInfo contextual information, such as location
+     * @param pushType the type of push notification
      */
-    private PushNotification(String mechanismUID, String messageId, String challenge, String amlbCookie,
-                               Calendar timeAdded, Calendar timeExpired, long ttl, boolean approved,
-                               boolean pending) {
+    private PushNotification(String mechanismUID, String messageId, String message, String challenge,
+                             String amlbCookie, Calendar timeAdded, Calendar timeExpired, long ttl,
+                             boolean approved, boolean pending, String customPayload,
+                             String numbersChallenge, String contextInfo, PushType pushType) {
         this.id = mechanismUID + "-" + timeAdded.getTimeInMillis();
         this.mechanismUID = mechanismUID;
         this.messageId = messageId;
+        this.message = message;
         this.challenge = challenge;
         this.amlbCookie = amlbCookie;
         this.timeAdded = timeAdded;
@@ -72,6 +97,10 @@ public class PushNotification extends ModelObject<PushNotification> {
         this.ttl = ttl;
         this.approved = approved;
         this.pending = pending;
+        this.customPayload = customPayload;
+        this.numbersChallenge = numbersChallenge;
+        this.contextInfo = contextInfo;
+        this.pushType = pushType;
     }
 
     /**
@@ -96,6 +125,14 @@ public class PushNotification extends ModelObject<PushNotification> {
      */
     public String getMessageId() {
         return messageId;
+    }
+
+    /**
+     * Get the message of this notification
+     * @return The message value
+     */
+    public String getMessage() {
+        return message;
     }
 
     /**
@@ -124,7 +161,7 @@ public class PushNotification extends ModelObject<PushNotification> {
 
     /**
      * Get the time that this notification was received.
-     * @return The date the notification was receuved.
+     * @return The date the notification was received.
      */
     public Calendar getTimeAdded() {
         return timeAdded;
@@ -136,6 +173,48 @@ public class PushNotification extends ModelObject<PushNotification> {
      */
     public Calendar getTimeExpired() {
         return timeExpired;
+    }
+
+    /**
+     * Get the contextual information to this notification.
+     * @return JSON String containing context information to this notification.
+     */
+    public String getContextInfo() {
+        return contextInfo;
+    }
+
+    /**
+     * Get the custom attributes added to this notification.
+     * @return JSON String containing custom attributes added to the payload of this notification.
+     */
+    public String getCustomPayload() {
+        return customPayload;
+    }
+
+    /**
+     * Get the type of Push notification.
+     * @return the push type.
+     */
+    public PushType getPushType() {
+        return pushType;
+    }
+
+    /**
+     * Get numbers used for push challenge
+     * @return the numbers as int array. Returns {null} if "numbersChallenge" is not available
+     */
+    public int[] getNumbersChallenge () {
+        int[] numbers = null;
+
+        if(this.numbersChallenge != null) {
+            String[] strArray = this.numbersChallenge.split(",");
+            numbers = new int[strArray.length];
+            for(int i = 0; i < strArray.length; i++) {
+                numbers[i] = Integer.parseInt(strArray[i]);
+            }
+        }
+
+        return numbers;
     }
 
     /**
@@ -200,6 +279,7 @@ public class PushNotification extends ModelObject<PushNotification> {
             jsonObject.put("id", getId());
             jsonObject.put("mechanismUID", getMechanismUID());
             jsonObject.put("messageId", getMessageId());
+            jsonObject.put("message", getMessage());
             jsonObject.put("challenge", getChallenge());
             jsonObject.put("amlbCookie", getAmlbCookie());
             jsonObject.put("timeAdded", getTimeAdded() != null ? getTimeAdded().getTimeInMillis() : null);
@@ -207,6 +287,12 @@ public class PushNotification extends ModelObject<PushNotification> {
             jsonObject.put("ttl", getTtl());
             jsonObject.put("approved", isApproved());
             jsonObject.put("pending", isPending());
+            jsonObject.put("customPayload", getCustomPayload());
+            jsonObject.put("numbersChallenge", this.numbersChallenge);
+            jsonObject.put("contextInfo", getContextInfo());
+            if(getPushType() != null) {
+                jsonObject.put("pushType", getPushType().toString());
+            }
         } catch (JSONException e) {
             Logger.warn(TAG, e, "Error parsing PushNotification object to JSON for messageId: %s",
                     messageId);
@@ -221,16 +307,79 @@ public class PushNotification extends ModelObject<PushNotification> {
     }
 
     /**
-     * Accepts the push authentication request.
-     * @param listener Listener for receiving the HTTP call response code.
+     * Accepts the push authentication request. Use this method to approve notification of
+     * type {@code PushType.DEFAULT}.
+     * @param listener Listener for receiving the authentication result.
      */
     public final void accept(@NonNull FRAListener<Void> listener) {
-        Logger.debug(TAG, "Accept Push Authentication request for message: %s", getMessageId());
-        performAcceptDenyAsync(true, listener);
+        if (this.pushType == PushType.DEFAULT) {
+            Logger.debug(TAG, "Accept Push Authentication request for message: %s", getMessageId());
+            performAcceptDenyAsync(true, listener);
+        } else {
+            listener.onException(new PushMechanismException("Error processing the Push " +
+                    "Authentication request. This method cannot be used to process notification " +
+                    "of type: " + this.pushType));
+        }
     }
 
     /**
-     * Deny the push authentication request.
+     * Accepts the push notification request with the challenge response. Use this method to handle
+     * notification of type {@code PushType.CHALLENGE}.
+     * @param challengeResponse the response for the Push Challenge
+     * @param listener Listener for receiving the authentication result.
+     */
+    public final void accept(@NonNull String challengeResponse, @NonNull FRAListener<Void> listener) {
+        if (this.pushType == PushType.CHALLENGE) {
+            Logger.debug(TAG, "Respond the challenge for message: %s", getMessageId());
+            PushResponder.getInstance().authenticationWithChallenge(this, challengeResponse, listener);
+        } else {
+            listener.onException(new PushMechanismException("Error processing the Push " +
+                    "Authentication request. This method cannot be used to process notification " +
+                    "of type: " + this.pushType));
+        }
+    }
+
+    /**
+     * Accepts the push notification request with Biometric Authentication. Use this method to handle
+     * notification of type {@code PushType.BIOMETRIC}.
+     * @param title the title to be displayed on the prompt.
+     * @param subtitle the subtitle to be displayed on the prompt.
+     * @param allowDeviceCredentials if {@code true}, accepts device PIN, pattern, or password to process notification.
+     * @param activity the activity of the client application that will host the prompt.
+     * @param listener listener for receiving the push authentication result.
+     */
+    @RequiresApi(Build.VERSION_CODES.M)
+    public final void accept(String title,
+                             String subtitle,
+                             boolean allowDeviceCredentials,
+                             @NonNull AppCompatActivity activity,
+                             @NonNull FRAListener<Void> listener) {
+        if (this.pushType == PushType.BIOMETRIC) {
+            final PushNotification pushNotification = this;
+            BiometricAuth biometricAuth = new BiometricAuth(title,
+                    subtitle, allowDeviceCredentials, activity, new BiometricAuthCompletionHandler() {
+                @Override
+                public void onSuccess(BiometricPrompt.AuthenticationResult result) {
+                    Logger.debug(TAG, "Respond the challenge for message: %s", getMessageId());
+                    PushResponder.getInstance().authentication(pushNotification, true, listener);
+                }
+
+                @Override
+                public void onError(int errorCode, String errorMessage) {
+                    listener.onException(new PushMechanismException("Error processing the Push " +
+                            "Authentication request. Biometric Authentication failed: " + errorMessage));
+                }
+            });
+            biometricAuth.authenticate();
+        } else {
+            listener.onException(new PushMechanismException("Error processing the Push " +
+                    "Authentication request. This method cannot be used to process notification " +
+                    "of type: " + this.pushType));
+        }
+    }
+
+    /**
+     * Deny any type of push authentication request.
      * @param listener Listener for receiving the HTTP call response code.
      */
     public final void deny(@NonNull FRAListener<Void> listener) {
@@ -258,6 +407,7 @@ public class PushNotification extends ModelObject<PushNotification> {
             return PushNotification.builder()
                     .setMechanismUID(jsonObject.getString("mechanismUID"))
                     .setMessageId(jsonObject.getString("messageId"))
+                    .setMessage(jsonObject.has("message") ? jsonObject.getString("message") : null)
                     .setChallenge(jsonObject.getString("challenge"))
                     .setAmlbCookie(jsonObject.has("amlbCookie") ? jsonObject.getString("amlbCookie") : null)
                     .setTimeAdded(jsonObject.has("timeAdded") ? getDate(jsonObject.optLong("timeAdded")) : null)
@@ -265,6 +415,10 @@ public class PushNotification extends ModelObject<PushNotification> {
                     .setTtl(jsonObject.optLong("ttl", -1))
                     .setApproved(jsonObject.getBoolean("approved"))
                     .setPending(jsonObject.getBoolean("pending"))
+                    .setCustomPayload(jsonObject.has("customPayload") ? jsonObject.getString("customPayload") : null)
+                    .setNumbersChallenge(jsonObject.has("numbersChallenge") ? jsonObject.getString("numbersChallenge") : null)
+                    .setContextInfo(jsonObject.has("contextInfo") ? jsonObject.getString("contextInfo") : null)
+                    .setPushType(jsonObject.has("pushType") ? jsonObject.getString("pushType") : PushType.DEFAULT.toString())
                     .build();
         } catch (JSONException | InvalidNotificationException e) {
             return null;
@@ -326,6 +480,7 @@ public class PushNotification extends ModelObject<PushNotification> {
     public static class PushNotificationBuilder {
         private String mechanismUID;
         private String messageId;
+        private String message;
         private String challenge;
         private String amlbCookie;
         private Calendar timeAdded;
@@ -333,6 +488,10 @@ public class PushNotification extends ModelObject<PushNotification> {
         private long ttl;
         private boolean approved = false;
         private boolean pending = true;
+        private String customPayload;
+        private String numbersChallenge;
+        private String contextInfo;
+        private PushType pushType;
         private Mechanism mechanism;
 
         /**
@@ -352,6 +511,16 @@ public class PushNotification extends ModelObject<PushNotification> {
          */
         public PushNotificationBuilder setMessageId(String messageId) {
             this.messageId = messageId;
+            return this;
+        }
+
+        /**
+         * Sets the message that was received with this notification.
+         * @param message The message that was received.
+         * @return The current builder.
+         */
+        public PushNotificationBuilder setMessage(String message) {
+            this.message = message;
             return this;
         }
 
@@ -426,6 +595,46 @@ public class PushNotification extends ModelObject<PushNotification> {
         }
 
         /**
+         * Sets custom attributes associated with this notification.
+         * @param customPayload JSON String containing the custom attributes
+         * @return The current builder.
+         */
+        public PushNotificationBuilder setCustomPayload(String customPayload) {
+            this.customPayload = customPayload;
+            return this;
+        }
+
+        /**
+         * Sets the numbers used in the Push challenge.
+         * @param numbersChallenge String containing the numbers for challenge
+         * @return The current builder.
+         */
+        public PushNotificationBuilder setNumbersChallenge(String numbersChallenge) {
+            this.numbersChallenge = numbersChallenge;
+            return this;
+        }
+
+        /**
+         * Sets context information for this notification.
+         * @param contextInfo JSON String containing the context information.
+         * @return The current builder.
+         */
+        public PushNotificationBuilder setContextInfo(String contextInfo) {
+            this.contextInfo = contextInfo;
+            return this;
+        }
+
+        /**
+         * Sets the type of Push notification.
+         * @param pushType the push type
+         * @return The current builder.
+         */
+        public PushNotificationBuilder setPushType(String pushType) {
+            this.pushType = PushType.fromString(pushType);
+            return this;
+        }
+
+        /**
          * Build the notification.
          * @return The final notification.
          * @throws InvalidNotificationException if timeAdded or mechanismUID are not provided
@@ -438,8 +647,8 @@ public class PushNotification extends ModelObject<PushNotification> {
                 throw new InvalidNotificationException("mechanismUID cannot be null.");
             }
 
-            return new PushNotification(mechanismUID, messageId, challenge, amlbCookie, timeAdded,
-                    timeExpired, ttl, approved, pending);
+            return new PushNotification(mechanismUID, messageId, message, challenge, amlbCookie,
+                    timeAdded, timeExpired, ttl, approved, pending, customPayload, numbersChallenge, contextInfo, pushType);
         }
     }